# docs/swagger.yaml
basePath: /
definitions:
  model.ErrorBody:
    properties:
      code:
        type: string
      message:
        type: string
      status:
        type: integer
    type: object
...
  model.Refund:
    properties:
      amount:
        type: number
      created_at:
        type: string
      decided_at:
        description: nullable
        type: string
      decided_by:
        description: nullable
        type: integer
      id:
        type: integer
      invoice_id:
        type: integer
      invoice_number:
        type: string
      merchant:
        type: string
      merchant_id:
        type: integer
      payment_intent_id:
        type: integer
      payment_intent_number:
        type: string
      processed_at:
        description: nullable
        type: string
      processed_by:
        description: nullable
        type: integer
      reason:
        type: string
      refund_number:
        type: string
      request_by:
        type: integer
      status:
        type: string
    type: object
  model.RefundDecideRequestBody:
    properties:
      status:
        enum:
        - APPROVE
        - REJECT
        example: APPROVE
        type: string
    type: object
  model.RefundRequest:
    properties:
      payment_intent_code:
        type: string
      reason:
        type: string
    type: object
  model.RefundResponse:
    properties:
      data:
        $ref: '#/definitions/model.Refund'
      error:
        $ref: '#/definitions/model.ErrorBody'
    type: object
...
  response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/model.ErrorBody'
    type: object
  response.SuccessResponse:
    properties:
      data: {}
      meta: {}
    type: object
host: localhost:8080
info:
  contact: {}
  description: API documentation for Project VioPal
  title: VioPal
  version: "1.0"
paths:
...
  /api/v1/admin/refunds:
    get:
      consumes:
      - application/json
      description: Get paginated refund list with optional filters (status)
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 10)
        in: query
        name: perpage
        type: integer
      - description: Invoice status (REQUESTED → APPROVED / REJECTED → SUCCESS / FAILED)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List refund (Admin)
      tags:
      - Admin
  /api/v1/admin/refunds/{refund_number}/decision:
    patch:
      consumes:
      - application/json
      description: Approve or reject a refund request. Only refunds with status REQUESTED
        can be decided.
      parameters:
      - description: Refund Number
        in: path
        name: refund_number
        required: true
        type: string
      - description: Refund decision payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.RefundDecideRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.RefundResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorBody'
      security:
      - BearerAuth: []
      summary: Decide refund (approve/reject)
      tags:
      - '[ADMIN] Refund'
  /api/v1/admin/refunds/{refund_number}/process:
    patch:
      consumes:
      - application/json
      description: Execute the refund process for a previously approved refund. This
        will move the refund to a final state (e.g., SUCCESS or FAILED).
      parameters:
      - description: Refund Number
        in: path
        name: refund_number
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.RefundResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorBody'
      security:
      - BearerAuth: []
      summary: Process refund
      tags:
      - '[ADMIN] Refund'
...
  /api/v1/refunds:
    get:
      consumes:
      - application/json
      description: Get paginated refund list with optional filters (status)
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 10)
        in: query
        name: perpage
        type: integer
      - description: Invoice status (REQUESTED → APPROVED / REJECTED → SUCCESS / FAILED)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List refund
      tags:
      - Refund
    post:
      consumes:
      - application/json
      description: Create a refund request based on payment intent code
      parameters:
      - description: Refund Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.RefundRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Request refund
      tags:
      - Refund
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"